refactor: replace TasksMutateDrawer with TasksMutateForm for task creation and editing
